add code

+++ boom-job-common/src/main/java/me/stevenkin/boom/job/common/dto/JobInstanceShardDto.java
@@ -15,8 +15,8 @@ public class JobInstanceShardDto implements Serializable {
     private String jobKey;
     private Long jobInstanceId;
     private Long jobShardId;
-    private Long jobShardIndex;
+    private Integer jobShardIndex;
     private String jobShardParam;
     private String jobParam;
-    private Long jobShardCount;
+    private Integer jobShardCount;
 }
